the-state-machine/python/requirements.txt
-e .
aws-cdk.core==1.119.0
aws-cdk.aws-lambda==1.119.0
aws-cdk.aws-stepfunctions==1.119.0
aws-cdk.aws-stepfunctions-tasks==1.119.0
aws-cdk.aws-iam==1.119.0
aws-cdk.aws_apigatewayv2==1.119.0
